requirements.txt
django
djangorestframework
asgiref
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
markdown 
django-filter
django-heroku
channels-redis 
gunicorn
altair                        
asgiref                       
attrs                         
autobahn                      
Automat                       
blinker                       
cachetools                    
certifi                       
cffi                          
channels                      
channels-redis                
charset-normalizer            
click                         
colorama                      
constantly                    
cryptography                  
daphne                        
dj-database-url               
Django                        
django-cors-headers           
django-environ               
django-filter                
django-heroku                
djangorestframework           
djangorestframework-simplejwt 
gitdb                         
GitPython                     
gunicorn                      
hyperlink                     
idna                          
incremental                   
Jinja2                        
jsonschema                    
jsonschema-specifications     
livekit                       
Markdown                      
markdown-it-py                

MarkupSafe                    
mdurl                         
msgpack                       
narwhals                      
numpy                         
packaging                     
pandas                        
pillow                        
pip                           
protobuf                      
psycopg2                      
psycopg2-binary               
pyarrow                       
pyasn1                        
pyasn1_modules                
pycparser                     
pydeck                        
Pygments                     
PyJWT                        
pyOpenSSL                   
python-dateutil               
python-dotenv                
pytz                         
redis                        
referencing                
requests                     
rich                          
rpds-py                      
ruff                        
service-identity           
setuptools                    
six                          
smmap                        
sqlparse                      
streamlit                    
tenacity                     
toml                        
tornado                      
Twisted                      
txaio                         
types-protobuf               
typing_extensions            
tzdata                       
urllib3                       
watchdog                     
whitenoise                    
zope.interface    
dj-database-url
psycopg2-binary           
